requirements/common.txt
Django==1.11.4
django-cors-headers==2.1.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
#PyJWT==1.5.2
#pytz==2017.2
#six==1.11.0
django-filter==1.0.4
python-sjsclient==0.8.0
pywebhdfs==0.4.1
Fabric==1.13.2
Fabric3
#watson-developer-cloud==0.26.1
gunicorn==19.7.1
django-redis==4.8.0
redis==3.2.0
mysqlclient==1.3.12
python-magic==0.4.13
simplejson==3.11.1
bs4==0.0.1
django-material==1.1.1
django-auditlog==0.4.4
yarn-api-client==0.2.3
fabric-gunicorn==0.1
matplotlib==2.2.4
knit==0.2.4
#pygments
celery==4.2.1
django-guardian==1.4.9
deprecation
newsapi-python
celery-once==2.0.0
psycopg2
django_celery_beat==1.4.0
django-environ
ls
boto3
ibm-cloud-sdk-core==1.5.1
ibm_watson==4.2.1
Pillow
